\name{local_extremum}
\alias{.local_min}
\alias{.local_max}
\title{
  Find Local Minima and Maxima in a Time-Series
}
\description{

  Find local minima and maxima in a time-series by comparing
  points with their neighbours.
  
}
\usage{
    pkg_build(pkg, parent.dir = ".", check = FALSE,
              build.vignettes = TRUE, 
              run.tests = TRUE, install = FALSE,
              clean = FALSE, bump.version = FALSE) 
    latest_version(pkg, path = ".") 
}
\arguments{
  \item{x}{
    a vector (or a list)
}
  \item{n}{
    an integer
}
  \item{first}{
    where to start
}
}
\details{

    \code{pkg_build}
}
\value{
  a vector (or a list)
}
\author{
  Enrico Schumann
}
\seealso{
  \code{\link{Extract}}
}
\examples{
1+1
}
